Update and extract combined-json plugins
