+++ .gitignore
@@ -8,4 +8,5 @@ typechain-types
 # Hardhat files
 cache
 artifacts
+contracts-artifacts
